feat:authorization

+++ src/module/contact/repository/prisma/contact-prisma.repository.ts
@@ -1,5 +1,9 @@
 import { plainToInstance } from 'class-transformer';
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  ConflictException,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/database/prisma.service';
 import { ContactRepository } from '../contact.repository';
 import { CreateContactDto } from '../../dto/create-contact.dto';
@@ -9,8 +13,7 @@ import { UpdateContactDto } from '../../dto/update-contact.dto';
 @Injectable()
 export class ContactPrismaRepository implements ContactRepository {
   constructor(private prisma: PrismaService) {}
-  async create(data: CreateContactDto): Promise<Contact> {
-    console.log(data);
+  async create(data: CreateContactDto, id: string): Promise<Contact> {
     const contact = new Contact();
 
     Object.assign(contact, { ...data });
@@ -23,49 +26,95 @@ export class ContactPrismaRepository implements ContactRepository {
         telephone: contact.telephone,
         created_at: contact.created_at,
         instagram: contact.instagram,
-        telegarm: contact.telegarm,
-        cleintId: contact.client_id,
+        telegram: contact.telegram,
+        clientId: id,
       },
     });
 
     return plainToInstance(Contact, newContact);
   }
 
-  async findAll(): Promise<Contact[]> {
-    const contact = await this.prisma.contact.findMany();
-    return plainToInstance(Contact, contact);
-  }
-
-  async findOne(id: string): Promise<Contact> {
-    const contact = await this.prisma.contact.findUnique({
-      where: { id },
+  async findAll(clientId: string): Promise<Contact[]> {
+    const contact = await this.prisma.contact.findMany({
+      where: {
+        clientId,
+      },
     });
-
     return plainToInstance(Contact, contact);
   }
 
-  async findByEmail(email: string): Promise<Contact> {
+  async findOne(id: string, clientId: string): Promise<Contact> {
     const contact = await this.prisma.contact.findFirst({
-      where: { email },
+      where: { id, clientId },
     });
 
-    return contact;
+    if (!contact) {
+      throw new NotFoundException('Contact not found!');
+    }
+
+    return plainToInstance(Contact, contact);
   }
 
-  async update(id: string, data: UpdateContactDto): Promise<Contact> {
+  async findByEmail(email: string, clientId: string): Promise<void> {
+    const contact = await this.prisma.contact.findFirst({
+      where: { clientId, email },
+    });
+
+    if (contact) {
+      throw new ConflictException('Contact alread exists!');
+    }
+  }
+
+  async update(
+    id: string,
+    data: UpdateContactDto,
+    clientId: string,
+  ): Promise<Contact> {
+    const checkClient = await this.prisma.contact.findFirst({
+      where: {
+        id: id,
+        clientId,
+      },
+    });
+
+    if (!checkClient) {
+      throw new NotFoundException('Contact found!');
+    }
+
     const contact = await this.prisma.contact.update({
       where: { id },
-      data: { ...data },
+      data: data,
     });
 
     return plainToInstance(Contact, contact);
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(id: string, clientId: string): Promise<void> {
+    const contact = await this.prisma.contact.findFirst({
+      where: {
+        id,
+        clientId,
+      },
+    });
+
+    if (!contact) {
+      throw new NotFoundException('Contact found!');
+    }
+
     await this.prisma.contact.delete({
       where: {
         id,
       },
     });
+  }
+
+  async findByTelephone(telephone: string, clientId: string): Promise<void> {
+    const contact = await this.prisma.contact.findFirst({
+      where: { telephone, clientId },
+    });
+
+    if (contact) {
+      throw new ConflictException('Telephone alread exists!');
+    }
   }
 }
